// clients/openapi-generator/typescript-node/model/apiStatusResponse.ts
/**
 * prompton-api
 * API for prompton - managing full lifecycle of AI chat prompts.
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { DBStatus } from './dBStatus';

export class ApiStatusResponse {
    'version': string;
    'message': string;
    'dbstatus': DBStatus;
    'githubSha'?: string;
    'githubEnv'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "version",
            "baseName": "version",
            "type": "string"
        },
        {
            "name": "message",
            "baseName": "message",
            "type": "string"
        },
        {
            "name": "dbstatus",
            "baseName": "dbstatus",
            "type": "DBStatus"
        },
        {
            "name": "githubSha",
            "baseName": "github_sha",
            "type": "string"
        },
        {
            "name": "githubEnv",
            "baseName": "github_env",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ApiStatusResponse.attributeTypeMap;
    }
}