// clients/openapi-generator/typescript-node/model/chatGPTChatCompletitionRequest.ts
/**
 * prompton-api
 * API for prompton - managing full lifecycle of AI chat prompts.
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ChatGPTMessage } from './chatGPTMessage';
import { Stop } from './stop';

export class ChatGPTChatCompletitionRequest {
    'model'?: string;
    'temperature'?: number;
    'topP'?: number;
    'stop'?: Stop;
    'maxTokens'?: number;
    'presencePenalty'?: number;
    'frequencyPenalty'?: number;
    'logitBias'?: { [key: string]: number; };
    'messages': Array<ChatGPTMessage>;
    'n'?: number = 1;
    'stream'?: boolean = false;
    'user'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "model",
            "baseName": "model",
            "type": "string"
        },
        {
            "name": "temperature",
            "baseName": "temperature",
            "type": "number"
        },
        {
            "name": "topP",
            "baseName": "top_p",
            "type": "number"
        },
        {
            "name": "stop",
            "baseName": "stop",
            "type": "Stop"
        },
        {
            "name": "maxTokens",
            "baseName": "max_tokens",
            "type": "number"
        },
        {
            "name": "presencePenalty",
            "baseName": "presence_penalty",
            "type": "number"
        },
        {
            "name": "frequencyPenalty",
            "baseName": "frequency_penalty",
            "type": "number"
        },
        {
            "name": "logitBias",
            "baseName": "logit_bias",
            "type": "{ [key: string]: number; }"
        },
        {
            "name": "messages",
            "baseName": "messages",
            "type": "Array<ChatGPTMessage>"
        },
        {
            "name": "n",
            "baseName": "n",
            "type": "number"
        },
        {
            "name": "stream",
            "baseName": "stream",
            "type": "boolean"
        },
        {
            "name": "user",
            "baseName": "user",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return ChatGPTChatCompletitionRequest.attributeTypeMap;
    }
}