// clients/openapi-generator/typescript-node/model/chatGPTCompletitionChoice.ts
/**
 * prompton-api
 * API for prompton - managing full lifecycle of AI chat prompts.
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ChatGPTMessage } from './chatGPTMessage';

export class ChatGPTCompletitionChoice {
    'message': ChatGPTMessage;
    'finishReason': string;
    'index': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "message",
            "baseName": "message",
            "type": "ChatGPTMessage"
        },
        {
            "name": "finishReason",
            "baseName": "finish_reason",
            "type": "string"
        },
        {
            "name": "index",
            "baseName": "index",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return ChatGPTCompletitionChoice.attributeTypeMap;
    }
}