// clients/openapi-generator/typescript-node/model/chatGPTTokenUsage.ts
/**
 * prompton-api
 * API for prompton - managing full lifecycle of AI chat prompts.
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class ChatGPTTokenUsage {
    'promptTokens': number;
    'completionTokens': number;
    'totalTokens': number;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "promptTokens",
            "baseName": "prompt_tokens",
            "type": "number"
        },
        {
            "name": "completionTokens",
            "baseName": "completion_tokens",
            "type": "number"
        },
        {
            "name": "totalTokens",
            "baseName": "total_tokens",
            "type": "number"
        }    ];

    static getAttributeTypeMap() {
        return ChatGPTTokenUsage.attributeTypeMap;
    }
}