// clients/openapi-generator/typescript-node/model/dBStatus.ts
/**
 * prompton-api
 * API for prompton - managing full lifecycle of AI chat prompts.
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class DBStatus {
    'statusCode': number;
    'statusMessage': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "statusCode",
            "baseName": "status_code",
            "type": "number"
        },
        {
            "name": "statusMessage",
            "baseName": "status_message",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return DBStatus.attributeTypeMap;
    }
}