// clients/openapi-generator/typescript-node/model/inferenceCreateByPromptVersionId.ts
/**
 * prompton-api
 * API for prompton - managing full lifecycle of AI chat prompts.
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';

export class InferenceCreateByPromptVersionId {
    /**
    * The API consumer\'s internal user reference for metrics. It is also relayed to the provider as part of the request if the provider supports it (eg. OpenAI\'s user field).
    */
    'endUserId'?: string;
    /**
    * The API consumer\'s source for metrics (e.g. AndroidApp etc).
    */
    'source'?: string;
    /**
    * The API consumer\'s internal reference id to able to link references to their sessions.
    */
    'clientRefId'?: string;
    'templateArgs'?: { [key: string]: string; };
    'metadata'?: object;
    /**
    * Provider request timout in seconds. If not provided, then Prompton API\'s default timeout for the provider will be used (90sec or `DEFAULT_OPENAI_REQUEST_TIMEOUT_SECONDS` env var if provided).
    */
    'requestTimeout'?: number;
    'promptVersionId': string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "endUserId",
            "baseName": "end_user_id",
            "type": "string"
        },
        {
            "name": "source",
            "baseName": "source",
            "type": "string"
        },
        {
            "name": "clientRefId",
            "baseName": "client_ref_id",
            "type": "string"
        },
        {
            "name": "templateArgs",
            "baseName": "template_args",
            "type": "{ [key: string]: string; }"
        },
        {
            "name": "metadata",
            "baseName": "metadata",
            "type": "object"
        },
        {
            "name": "requestTimeout",
            "baseName": "request_timeout",
            "type": "number"
        },
        {
            "name": "promptVersionId",
            "baseName": "prompt_version_id",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return InferenceCreateByPromptVersionId.attributeTypeMap;
    }
}