// clients/openapi-generator/typescript-node/model/inferencePostResponse.ts
/**
 * prompton-api
 * API for prompton - managing full lifecycle of AI chat prompts.
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { Response } from './response';

/**
* Used for FastAPI response model
*/
export class InferencePostResponse {
    'id': string;
    'response': Response;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "id",
            "baseName": "id",
            "type": "string"
        },
        {
            "name": "response",
            "baseName": "response",
            "type": "Response"
        }    ];

    static getAttributeTypeMap() {
        return InferencePostResponse.attributeTypeMap;
    }
}