// clients/openapi-generator/typescript-node/model/inferenceRequestData.ts
/**
 * prompton-api
 * API for prompton - managing full lifecycle of AI chat prompts.
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { RequestFile } from './models';
import { ChatGPTChatCompletitionRequest } from './chatGPTChatCompletitionRequest';
import { PromptVersionProviders } from './promptVersionProviders';

export class InferenceRequestData {
    'provider': PromptVersionProviders;
    'rawRequest': ChatGPTChatCompletitionRequest;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "provider",
            "baseName": "provider",
            "type": "PromptVersionProviders"
        },
        {
            "name": "rawRequest",
            "baseName": "raw_request",
            "type": "ChatGPTChatCompletitionRequest"
        }    ];

    static getAttributeTypeMap() {
        return InferenceRequestData.attributeTypeMap;
    }
}

export namespace InferenceRequestData {
}